+++ Data.lua
@@ -5,7 +5,7 @@ local Data = DKM.Data
 local T = DKM.T or function(value) return value end
 
 Data.addonName = ADDON_NAME
-Data.version = "1.1.9"
+Data.version = "1.1.10"
 Data.interface = 120100
 Data.dataVersion = "2026-08-23"
 Data.patch = "12.1.0"